Fix for #212 updatebot fails

+++ chatbot-conversational_AI/update-bot/app/update-bot.js
@@ -10,12 +10,14 @@ async function main() {
   var update_all_questions = false;
   if ('UPDATE_ALL' in process.env && process.env.UPDATE_ALL == 'Y') {
     update_all_questions = true;
+    console.log("Updating all questions.");
   }
   console.log("********************************** all knowledge received **********************************")
 
   for (let counter = 0; counter < stackData.items.length; counter++) {
     // Choose the next question from Stack
     stackQuestion = stackData.items[counter];
+    var stack_q_id = stackQuestion.question_id;
 
     // Check if the question is already in the database
     var q_in_db_flag = 0;
@@ -27,32 +29,38 @@ async function main() {
         break;
       }
     }
+    if (q_in_db_flag == 1) {
+      console.log(`${new Date().toISOString()}: Question (ID=${stack_q_id}) was found in mssql db at index ${id_in_db}`);
+    }
+    else {
+      console.log(`${new Date().toISOString()}: Question (ID=${stack_q_id}) was not found in mssql db.`);
+    }
 
     // If (question is not in database but answered) or all questions should be updated
-    if ((q_in_db_flag == 0 && stackQuestion.is_answered) || update_all_questions) {
-      var answers = await get_stackAnswers(stackQuestion.question_id),
-        questionText = stackQuestion.title;
-      questionLink = stackQuestion.link;
-      var bestAnswer = answers.items[0],
-        answerText = bestAnswer.body;
+    if ((q_in_db_flag == 0 && stackQuestion.is_answered) || (update_all_questions && stackQuestion.is_answered)) {
+      var answers = await get_stackAnswers(stackQuestion.question_id);
+      var questionText = stackQuestion.title;
+      var questionLink = stackQuestion.link;
+      var bestAnswer = answers.items[0];
+      var answerText = bestAnswer.body;
 
-      var stack_q_id = stackQuestion.question_id,
-        stack_q_ts = stackQuestion.last_activity_date
-      stack_a_id = bestAnswer.answer_id,
-        stack_a_ts = bestAnswer.last_edit_date ? bestAnswer.last_edit_date : bestAnswer.creation_date
-      cai_q_id = null,
-        cai_a_id = null;
+      var stack_q_ts = stackQuestion.last_activity_date;
+      var stack_a_id = bestAnswer.answer_id;
+      var stack_a_ts = bestAnswer.last_edit_date ? bestAnswer.last_edit_date : bestAnswer.creation_date;
+      var cai_q_id = null;
+      var cai_a_id = null;
 
 
       // Question is already in the database and needs to be updated (tested on 20210910)
       var q_update_flag = 0;
       if (q_in_db_flag == 1 && (dbData[id_in_db].stack_q_ts != stack_q_ts || dbData[id_in_db].stack_a_ts != stack_a_ts)) {
+        console.log(`${new Date().toISOString()}: Deleting Question (ID=${dbData[id_in_db].id_num}) from CAI and mssql, because it is already in the dbs but was updated (or received new answer).`);
         try { // Delete it from the database and CAI
           await delete_caiEntry(dbData[id_in_db].cai_a_id, caiCredentials.access_token);
           await db_request.query(`delete from Questions where id_num = '${dbData[id_in_db].id_num}'; `);
           q_update_flag = 1;
         } catch (err) {
-          console.log("An Error has occurred during deleting a Question/Answer pair in SAP CAI and the internal DB");
+          console.log(`${new Date().toISOString()}: An Error has occurred during deleting a Question/Answer pair (CAI_ANSWER_ID=${dbData[id_in_db].cai_a_id}; MSSQL_ID=${dbData[id_in_db].id_num}) in SAP CAI and the internal DB`);
           throw err;
         }
       }
@@ -60,17 +68,21 @@ async function main() {
       // Question is not in database or must be updated (was deleted and must be added again)
       if (q_in_db_flag == 0 || q_update_flag == 1) {
         try {
+          console.log(`${new Date().toISOString()}: Add answer to CAI (${answerText.substring(0, 50)})`);
           var arrTuple = await add_caiAnswer(answerText, questionLink, caiCredentials.access_token);
           var caiAnswerResult = arrTuple[0];
           var errValue = arrTuple[1];
+          console.log("Answer from CAI: " + caiAnswerResult);
           if (errValue === null) {
             cai_a_id = caiAnswerResult.results.id;
+            console.log(`${new Date().toISOString()}: Add question (for answer ID=${cai_a_id}) to CAI (${questionText})`);
             var caiQuestionResult = await add_caiQuestion(questionText, cai_a_id, caiCredentials.access_token);
             cai_q_id = caiQuestionResult.results.id;
+            console.log("Insert entry in mssql: stack_q_id=${stack_q_id}, stack_a_id=${stack_a_id}, cai_q_id=${cai_q_id}, cai_a_id=${cai_a_id}");
             await db_request.query(`insert into Questions (stack_q_id, stack_q_ts, stack_a_id, stack_a_ts, cai_q_id, cai_a_id) values ('${stack_q_id}', '${stack_q_ts}', '${stack_a_id}', '${stack_a_ts}', '${cai_q_id}', '${cai_a_id}'); select * from Questions where stack_q_id = '${stack_q_id}'`);
           }
         } catch (err) {
-          console.log("An Error has occurred during adding a new Question/Answer pair to SAP CAI and the internal DB");
+          console.log(`${new Date().toISOString()}: An Error has occurred during adding a new Question/Answer pair to SAP CAI and the internal DB`);
           throw err;
         }
       }
@@ -81,15 +93,15 @@ async function main() {
         await delete_caiEntry(dbData[id_in_db].cai_a_id, caiCredentials.access_token);
         await db_request.query(`delete from Questions where id_num = '${dbData[id_in_db].id_num}'`);
       } catch (err) {
-        console.log("An Error has occurred during deleting a Question/Answer pair in SAP CAI and the internal DB (it was deleted because it was deleted in Stack Overflow)");
+        console.log(`${new Date().toISOString()}: An Error has occurred during deleting a Question/Answer pair (CAI_ANSWER_ID=${dbData[id_in_db].cai_a_id}; MSSQL_ID=${dbData[id_in_db].id_num}) in SAP CAI and the internal DB (it was deleted because it was deleted in Stack Overflow)`);
         throw err;
       }
-      console.log("A question was deleted from Stack Overflow and hence in the bot.");
+      console.log(`${new Date().toISOString()}: A question (CAI_ANSWER_ID=${dbData[id_in_db].cai_a_id}; MSSQL_ID=${dbData[id_in_db].id_num}) was deleted from Stack Overflow and hence in the bot.`);
     }
   }
 
   await sqlconnection.close();
-  console.log("done");
+  console.log(`${new Date().toISOString()}: done`);
   process.exit(0);
 }
 
@@ -106,21 +118,21 @@ async function get_stackQuestions() {
     // An API call can fetch max 100 entries per page. It must be checked if there is more data available (see: https://api.stackexchange.com/docs/paging)
     var pagenumber = 1;
     var stack_url_questions = process.env.STACK_URL + '/search/advanced?tagged=' + process.env.STACK_TAG + '&pagesize=100&page=' + pagenumber + '&filter=withbody&key=' + process.env.STACK_KEY;
-    var result = await got(stack_url_questions, stack_config);
+    var result = await stack_request(stack_url_questions, stack_config);
     var allQuestions = result.body;
     var moreDataAvailable_flag = allQuestions.has_more;
 
     while (moreDataAvailable_flag) {
       pagenumber = pagenumber + 1;
       stack_url_questions = process.env.STACK_URL + '/search/advanced?tagged=' + process.env.STACK_TAG + '&pagesize=100&page=' + pagenumber + '&filter=withbody&key=' + process.env.STACK_KEY;
-      result = await got(stack_url_questions, stack_config);
+      result = await stack_request(stack_url_questions, stack_config);
       result.body.items.forEach(element => allQuestions.items.push(element));
       moreDataAvailable_flag = result.body.has_more;
     }
 
     return allQuestions;
   } catch (err) {
-    console.log("An Error has occurred during requesting the stack questions labeled with " + process.env.STACK_TAG + ". Maybe it is a problem with concatenating multiple pages of questions because max pagesize exceeded.");
+    console.log(`${new Date().toISOString()}: An Error has occurred during requesting the stack questions labeled with ${process.env.STACK_TAG}. Maybe it is a problem with concatenating multiple pages of questions because max pagesize exceeded.`);
     throw err;
   }
 }
@@ -128,10 +140,31 @@ async function get_stackQuestions() {
 async function get_stackAnswers(question_id) {
   const stack_url_answers = process.env.STACK_URL + '/questions/' + question_id + '/answers?pagesize=100&filter=withbody&sort=votes&key=' + process.env.STACK_KEY;
   try {
-    const result = await got(stack_url_answers, stack_config);
+    const result = await stack_request(stack_url_answers, stack_config);
     return result.body;
   } catch (err) {
-    console.log("An Error has occurred during requesting the stack answer to question " + question_id);
+    console.log(`${new Date().toISOString()}: An Error has occurred during requesting the stack answer to question ${question_id}`);
+    throw err;
+  }
+}
+
+async function stack_request(url, config) {
+  try {
+    console.log(`${new Date().toISOString()}: STACK_REQUEST: ${url}`);
+    const result = await got(url, config);
+    // application must wait some seconds if backoff-field in response is set
+    // otherwise a throttle-violation is raised
+    if ("backoff" in result.body) {
+      var backoff_seconds = result.body["backoff"];
+      console.log(`${new Date().toISOString()}: Backoff response received from Stack! Duration: ${backoff_seconds} seconds`);
+      // add some offset (2 seconds) to be sure that the backoff is long enough.
+      // the stack server does not measure the time differences very accurately (and not in favor of the client)
+      await new Promise(r => setTimeout(r, (backoff_seconds+2) * 1000));
+      console.log(`${new Date().toISOString()}: Backoff finsihed!`);
+    }
+    return result;
+  } catch(err) {
+    console.log(`${new Date().toISOString()}: Error in stack_request!\nURL: ${url}\nError: ${err}`);
     throw err;
   }
 }
@@ -212,7 +245,8 @@ async function add_caiAnswer(answerText, questionLink, access_token) {
 
     // add the Link to Stack Overflow to the Answer
     if (slackFormat.length > 1800) { // answer is too long
-      finalFormat = "\n" + slackFormat.substring(0, slackFormat.split('. ', 4).join('. ').length + 1) + " ... " + " <" + questionLink + "/|[See more]>";
+      // cut the answer to max four sentences and max 1800 characters
+      finalFormat = "\n" + slackFormat.substring(0, Math.min(slackFormat.split('. ', 4).join('. ').length + 1, 1800)) + " ... " + " <" + questionLink + "/|[See more]>";
     } else {
       var finalFormat = "\n" + slackFormat + "\n\n" + "For more help, please click <" + questionLink + "/|here> to go directly to this question on Stack Overflow.";
     }
@@ -226,6 +260,7 @@ async function add_caiAnswer(answerText, questionLink, access_token) {
     return [JSON.parse(result.body), null];
   } catch (err) {
     console.log("An Error has occurred during adding an answer to SAP CAI: " + err);
+    console.log(`cai_request_url=${cai_request_url}; finalAnswerString=${finalAnswerString}`);
     //throw err;
     return [null, err];
   }
